Implement new design for the profile page
