+++ frontend/src/locale/index.ts
@@ -34,12 +34,11 @@ export const messages = {
     "profile.loading": "Loading...",
     "profile.profileHeader": "My Profile",
     "profile.mySubmissionsHeader": "My Submissions",
-    "profile.editProfile":
-      "Something doesn't look right? Click here to change it!",
+    "profile.editProfile": "Edit your profile",
     "profile.myOrders": "My Orders",
     "profile.orderId": "Order #",
-    "profile.status": "Status: ",
-    "profile.price": "Price €",
+    "profile.status": "Status",
+    "profile.price": "Price",
     "profile.manageOrder": "Manage order",
     "profile.items": "Tickets",
     "profile.ticketFor": "Ticket for",
@@ -425,10 +424,10 @@ export const messages = {
     "profile.header": "Il mio profilo",
     "profile.loading": "Caricamento...",
     "profile.profileHeader": "Il mio profilo",
-    "profile.editProfile": "Informazioni errate? Clicca qui per cambiarle!",
+    "profile.editProfile": "Modifica profilo",
     "profile.orderId": "Ordine #",
-    "profile.status": "Stato: ",
-    "profile.price": "Prezzo €",
+    "profile.status": "Stato",
+    "profile.price": "Prezzo",
     "profile.items": "Biglietti",
     "profile.manageOrder": "Gestisci ordine",
     "profile.mySubmissionsHeader": "Le mie submissions",
