+++ frontend/src/pages/[lang]/profile/index.tsx
@@ -5,13 +5,13 @@ import { Fragment, useEffect } from "react";
 import { FormattedMessage } from "react-intl";
 import { jsx } from "theme-ui";
 
-import { useLoginState } from "~/app/profile/hooks";
-import { Logout } from "~/app/profile/logout";
-import { MyOrders } from "~/app/profile/my-orders";
-import { MyProfile } from "~/app/profile/my-profile";
-import { MySubmissions } from "~/app/profile/my-submissions";
 import { MetaTags } from "~/components/meta-tags";
 import { PageLoading } from "~/components/page-loading";
+import { useLoginState } from "~/components/profile/hooks";
+import { Logout } from "~/components/profile/logout";
+import { MyOrders } from "~/components/profile/my-orders";
+import { MyProfile } from "~/components/profile/my-profile";
+import { MySubmissions } from "~/components/profile/my-submissions";
 import { useCurrentLanguage } from "~/locale/context";
 import { useMyProfileQuery } from "~/types";
